Correccion del precio en detalle

+++ app/src/main/java/com/example/proyectolagranja/ui/Clases/ArticuloDetalle.java
@@ -3,11 +3,11 @@ package com.example.proyectolagranja.ui.Clases;
 public class ArticuloDetalle {
     private String nombre;
     private double precio;
-    private int cantidad;
+    private double cantidad;
     private double subTotal;
     private String imagenUrl;
 
-    public ArticuloDetalle(String nombre, double precio, int cantidad, double subTotal, String imagenUrl) {
+    public ArticuloDetalle(String nombre, double precio, double cantidad, double subTotal, String imagenUrl) {
         this.nombre = nombre;
         this.precio = precio;
         this.cantidad = cantidad;
@@ -17,7 +17,7 @@ public class ArticuloDetalle {
 
     public String getNombre() { return nombre; }
     public double getPrecio() { return precio; }
-    public int getCantidad() { return cantidad; }
+    public double getCantidad() { return cantidad; }
     public double getSubTotal() { return subTotal; }
     public String getImagenUrl() { return imagenUrl; }
 }
